% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_existing_data.R
\name{check_existing_data}
\alias{check_existing_data}
\title{Check and filter out existing data already present in the database}
\usage{
check_existing_data(con, schema = Sys.getenv("PG_SCHEMA"), new_data)
}
\arguments{
\item{con}{A valid DBI database connection.}

\item{schema}{The schema name (e.g., \"student_paul\").}

\item{new_data}{A tibble containing the new data to check, must include \code{index_ts} and \code{date} columns.}
}
\value{
A tibble containing only the new rows not already present in the database.
}
\description{
This function removes rows from a new dataset if they already exist in the target PostgreSQL table.
It avoids inserting duplicate entries based on a unique key (symbol + date or index_ts + date).
}
