% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{Format wide Yahoo Finance data into long format}
\usage{
format_data(data)
}
\arguments{
\item{data}{A tibble with columns like open, high, low, close, volume, close_adjusted, etc.}
}
\value{
A tibble in long format with columns: index_ts, date, metric, value.
}
\description{
This function reshapes a wide tibble of stock prices into a long format
compatible with the \code{data_sp500} database table.
}
