% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_new_data.R
\name{insert_new_data}
\alias{insert_new_data}
\title{Insert new data into the student's PostgreSQL table}
\usage{
insert_new_data(con, schema = Sys.getenv("PG_SCHEMA"), new_data)
}
\arguments{
\item{con}{A valid DBI database connection.}

\item{schema}{The schema name (e.g., \"student_paul\").}

\item{new_data}{A tibble containing the new rows to insert. Must match the database structure.}
}
\value{
True if all is well, otherwise an error is raised.
}
\description{
This function inserts new rows into the \code{data_sp500} table in the specified schema.
It assumes that the data has already been filtered by \code{check_existing_data()}.
}
