% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_summary_table.R
\name{push_summary_table}
\alias{push_summary_table}
\title{Push the summary table to PostgreSQL}
\usage{
push_summary_table(con, summary_table, user_login = Sys.getenv("user_login"))
}
\arguments{
\item{con}{A valid DBI database connection.}

\item{summary_table}{The tibble created during the pipeline (via build_summary_table and log_summary).}

\item{user_login}{Optional, defaults to Sys.getenv('user_login'). The login of the student (e.g., "paul", "fariba").}
}
\value{
Nothing. Pushes logs into the database.
}
\description{
This function writes the batch summary to the student's \code{pipeline_logs} table inside their schema.
}
