% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_pipeline.R
\name{start_pipeline}
\alias{start_pipeline}
\title{Start the full data pipeline}
\usage{
start_pipeline(from = Sys.Date() - 7, to = Sys.Date(), batch_size = 25)
}
\arguments{
\item{from}{A Date indicating the start date for historical data (default: 7 days ago).}

\item{to}{A Date indicating the end date for historical data (default: today).}

\item{batch_size}{An integer indicating how many symbols to fetch per batch (default: 25).}
}
\value{
Nothing. Runs the full data pipeline.
}
\description{
This function connects to the database, fetches the list of symbols,
downloads new OHLCV data from Yahoo Finance in batches,
filters out already existing data, inserts the new data into the database,
and logs each batch processing result.
}
